Move lutimes polyfills to separate file
