+++ package.json
@@ -46,6 +46,7 @@
     "check-for-callback.js",
     "clone.js",
     "graceful-fs.js",
+    "lutimes-polyfill.js",
     "noop.js",
     "normalize-args.js",
     "polyfills.js",
