+++ polyfills.js
@@ -28,7 +28,7 @@ function patch (fs) {
 
   // lutimes implementation, or no-op
   if (!fs.lutimes) {
-    patchLutimes(fs)
+    require('./lutimes-polyfill.js')(fs)
   }
 
   // https://github.com/isaacs/node-graceful-fs/issues/4
@@ -108,47 +108,6 @@ function patch (fs) {
 
         throw er
       }
-    }
-  }
-
-  function patchLutimes (fs) {
-    if (constants.hasOwnProperty("O_SYMLINK")) {
-      fs.lutimes = function (path, at, mt, cb) {
-        fs.open(path, constants.O_SYMLINK, function (er, fd) {
-          if (er) {
-            if (cb) cb(er)
-            return
-          }
-          fs.futimes(fd, at, mt, function (er) {
-            fs.close(fd, function (er2) {
-              if (cb) cb(er || er2)
-            })
-          })
-        })
-      }
-
-      fs.lutimesSync = function (path, at, mt) {
-        var fd = fs.openSync(path, constants.O_SYMLINK)
-        var ret
-        var threw = true
-        try {
-          ret = fs.futimesSync(fd, at, mt)
-          threw = false
-        } finally {
-          if (threw) {
-            try {
-              fs.closeSync(fd)
-            } catch (er) {}
-          } else {
-            fs.closeSync(fd)
-          }
-        }
-        return ret
-      }
-
-    } else {
-      fs.lutimes = noop
-      fs.lutimesSync = noopSync
     }
   }
